rename other related files
