+++ apps/teams-test-app/src/App.tsx
@@ -42,7 +42,7 @@ import MonetizationAPIs from './components/privateApis/MonetizationAPIs';
 import NotificationAPIs from './components/privateApis/NotificationAPIs';
 import PrivateAPIs from './components/privateApis/PrivateAPIs';
 import TeamsAPIs from './components/privateApis/TeamsAPIs';
-import VideoExAPIs from './components/privateApis/VideoExAPIs';
+import VideoExAPIs from './components/privateApis/VideoEffectsExAPIs';
 import ProfileAPIs from './components/ProfileAPIs';
 import RemoteCameraAPIs from './components/RemoteCameraAPIs';
 import SearchAPIs from './components/SearchAPIs';
@@ -52,7 +52,7 @@ import StageViewAPIs from './components/StageViewAPIs';
 import TeamsCoreAPIs from './components/TeamsCoreAPIs';
 import { isTestBackCompat } from './components/utils/isTestBackCompat';
 import Version from './components/Version';
-import VideoAPIs from './components/VideoApis';
+import VideoAPIs from './components/VideoEffectsApis';
 import WebStorageAPIs from './components/WebStorageAPIs';
 
 const urlParams = new URLSearchParams(window.location.search);
